+++ nicfit/logger.py
@@ -6,8 +6,6 @@ import logging.config
 from io import StringIO
 from textwrap import dedent
 
-from deprecation import deprecated
-from .__about__ import __version__
 from .config import Config
 
 __all__ = ["stdout", "stderr", "FileConfig", "DictConfig"]
@@ -328,52 +326,6 @@ format = "%(message)s"
         self.write(out)
         out.seek(0)
         return out.read()
-
-
-@deprecated(details="Use FileConfig.DEFAULT_LOGGING_CONFIG",
-            deprecated_in="0.6.3", removed_in="0.7",
-            current_version=__version__)
-def LOGGING_CONFIG(pkg_logger, root_level="WARN", log_format=LOG_FORMAT,
-                   pkg_level="NOTSET", init_logging=False):
-    cfg = """
-###
-#logging configuration
-#https://docs.python.org/3/library/logging.config.html#configuration-file-format
-###
-
-[loggers]
-keys = root, {pkg_logger}
-
-[handlers]
-keys = console
-
-[formatters]
-keys = generic
-
-[logger_root]
-level = {root_level}
-handlers = console
-
-[logger_{pkg_logger}]
-level = {pkg_level}
-qualname = {pkg_logger}
-; When adding more specific handlers than what exists on the root you'll
-; likely want to set propagate to false.
-handlers =
-propagate = 1
-
-[handler_console]
-class = StreamHandler
-args = (sys.stderr,)
-level = NOTSET
-formatter = generic
-
-[formatter_generic]
-format = {log_format}
-""".format(**locals())
-    if init_logging:
-        logging.config.fileConfig(StringIO(cfg))
-    return cfg
 
 
 if __name__ == "__main__":                                     # pragma: nocover